seperate sharing into own component
